+++ Modules/booking/Model/BkRestrictions.php
@@ -11,6 +11,10 @@ class BkRestrictions extends Model
         $this->setColumnsInfo("id_resource", "int(11)", 0);
         $this->setColumnsInfo("maxbookingperday", "int(11)", 0);
         $this->setColumnsInfo("bookingdelayusercanedit", "int(11)", 0);
+        $this->setColumnsInfo("maxduration", "varchar(50)", "");  // maximum booking time 5d, 10h
+        $this->setColumnsInfo("maxfullduration", "tinyint", 0);  // max real booking time (no closed hours) or count everything (duration between start and end time)
+        $this->setColumnsInfo("disableoverclosed", "tinyint", 0); // (dis)allow booking on multiple ranges with closed hours between
+        $this->setColumnsInfo("applies_to", "int", CoreSpace::$USER); // max role restrictions applies to
 
         $this->primaryKey = "id";
     }
@@ -21,8 +25,30 @@ class BkRestrictions extends Model
         $resources = $this->runRequest($sql, array($id_space))->fetchAll();
 
         foreach ($resources as $r) {
-            $this->add($id_space, $r["id"], 0, 0);
+            $this->add($id_space, $r["id"], 0, 0, '', 0, 0, CoreSpace::$USER);
         }
+    }
+
+    public function default($id_space, $id_resource)
+    {
+        return  [
+            'id' => 0,
+            'id_resource' => $id_resource,
+            'id_space' => $id_space,
+            'maxbookingperday' => 0,
+            'bookingdelayusercanedit' => 0,
+            'maxduration' => 0,
+            'maxfullduration' => 0,
+            'disableoverclosed' => 0,
+            'applies_to' => CoreSpace::$USER
+        ];
+    }
+
+    public function getForResource($id_space, $id_resource)
+    {
+        $sql = "SELECT * FROM bk_restrictions WHERE id_resource=? AND deleted=0 AND id_space=?";
+        $tmp = $this->runRequest($sql, array($id_resource, $id_space))->fetch();
+        return $tmp ? $tmp : $this->default($id_space, $id_resource);
     }
 
     public function getBookingDelayUserCanEdit($id_space, $id_resource)
@@ -51,19 +77,26 @@ class BkRestrictions extends Model
         return $this->runRequest($sql, array($id, $id_space))->fetch();
     }
 
-    public function add($id_space, $id_resource, $maxbookingperday, $bookingdelayusercanedit)
+    public function add($id_space, $id_resource, $maxbookingperday, $bookingdelayusercanedit, $maxduration, $maxfullduration, $disableoverclosed, $appliesTo)
     {
         $id = $this->exists($id_space, $id_resource);
         if (!$id) {
-            $sql = "INSERT INTO bk_restrictions (id_resource, maxbookingperday, bookingdelayusercanedit, id_space) VALUES (?,?,?,?)";
-            $this->runRequest($sql, array($id_resource, $maxbookingperday, $bookingdelayusercanedit, $id_space));
+            $sql = "INSERT INTO bk_restrictions (id_resource, maxbookingperday, bookingdelayusercanedit, id_space, maxduration, maxfullduration, disableoverclosed, applies_to) VALUES (?,?,?,?, ?, ?, ?, ?)";
+            $this->runRequest($sql, array($id_resource, $maxbookingperday, $bookingdelayusercanedit, $id_space, $maxduration, $maxfullduration, $disableoverclosed, $appliesTo));
+            $id = $this->getDatabase()->lastInsertId();
         }
+        return $id;
     }
 
-    public function set($id_space, $id, $maxbookingperday, $bookingdelayusercanedit)
+    public function set($id_space, $id, $id_resource, $maxbookingperday, $bookingdelayusercanedit, $maxduration, $maxfullduration, $disableoverclosed, $appliesTo)
     {
-        $sql = "UPDATE bk_restrictions SET maxbookingperday=?, bookingdelayusercanedit=? WHERE id=? AND deleted=0 AND id_space=?";
-        $this->runRequest($sql, array($maxbookingperday, $bookingdelayusercanedit, $id, $id_space));
+        if ($id) {
+            $sql = "UPDATE bk_restrictions SET id_resource=?, maxbookingperday=?, bookingdelayusercanedit=?, maxduration=?, maxfullduration=?, disableoverclosed=?, applies_to=? WHERE id=? AND deleted=0 AND id_space=?";
+            $this->runRequest($sql, array($id_resource, $maxbookingperday, $bookingdelayusercanedit, $maxduration, $maxfullduration, $disableoverclosed, $appliesTo, $id, $id_space));
+        } else {
+            $id = $this->add($id_space, $id_resource, $maxbookingperday, $bookingdelayusercanedit, $maxduration, $maxfullduration, $disableoverclosed, $appliesTo);
+        }
+        return $id;
     }
 
     public function exists($id_space, $id_resource)